// src/types/code.type.ts
export type Code = {
    codeKbn: string;
    code1: string;
    code2: string;
    code3:  string;
    content1: string;
    content2: string;
    content3: string;
    content4: string;
    content5: string;
    content6: string;
    content7: string;
    content8: string;
    content9: string;
}